feat: make web summary available behind ff
