+++ front/lib/actions/mcp_internal_actions/constants.test.ts
@@ -49,10 +49,6 @@ describe("INTERNAL_MCP_SERVERS", () => {
       { name: "content_creation", id: 23 },
       { name: "slideshow", id: 28 },
       { name: "deep_research", id: 29 },
-      {
-        name: "web_search_&_browse_with_summary",
-        id: 30,
-      },
       { name: "search", id: 1006 },
       { name: "run_agent", id: 1008 },
       { name: "reasoning", id: 1007 },
