+++ front/lib/actions/mcp_internal_actions/constants.ts
@@ -106,7 +106,6 @@ export const AVAILABLE_INTERNAL_MCP_SERVER_NAMES = [
   "think",
   "toolsets",
   "web_search_&_browse",
-  "web_search_&_browse_with_summary",
   SEARCH_SERVER_NAME,
   TABLE_QUERY_V2_SERVER_NAME,
 ] as const;
@@ -942,25 +941,6 @@ The directive should be used to display a clickable version of the agent name in
       icon: "ActionAtomIcon",
       documentationUrl: null,
       instructions: `This tool performs a complete handoff to the dust-deep research agent: ${DUST_DEEP_DESCRIPTION}`,
-    },
-  },
-  "web_search_&_browse_with_summary": {
-    id: 30,
-    availability: "auto_hidden_builder",
-    allowMultipleInstances: false,
-    isRestricted: undefined,
-    isPreview: true,
-    tools_stakes: undefined,
-    tools_retry_policies: { default: "retry_on_interrupt" },
-    timeoutMs: undefined,
-    serverInfo: {
-      name: "web_search_&_browse_with_summary",
-      version: "1.0.0",
-      description: DEFAULT_WEBSEARCH_ACTION_DESCRIPTION,
-      icon: "ActionGlobeAltIcon",
-      authorization: null,
-      documentationUrl: null,
-      instructions: null,
     },
   },
   slack_bot: {
