+++ front/lib/actions/types/guards.ts
@@ -71,14 +71,7 @@ export function isMCPConfigurationForInternalWebsearch(
 ): arg is ServerSideMCPServerConfigurationType {
   return (
     isServerSideMCPServerConfiguration(arg) &&
-    (isInternalMCPServerOfName(
-      arg.internalMCPServerId,
-      "web_search_&_browse"
-    ) ||
-      isInternalMCPServerOfName(
-        arg.internalMCPServerId,
-        "web_search_&_browse_with_summary"
-      ))
+    isInternalMCPServerOfName(arg.internalMCPServerId, "web_search_&_browse")
   );
 }
 
@@ -175,14 +168,7 @@ export function isMCPInternalWebsearch(
 ): arg is ServerSideMCPToolConfigurationType {
   return (
     isServerSideMCPToolConfiguration(arg) &&
-    (isInternalMCPServerOfName(
-      arg.internalMCPServerId,
-      "web_search_&_browse"
-    ) ||
-      isInternalMCPServerOfName(
-        arg.internalMCPServerId,
-        "web_search_&_browse_with_summary"
-      ))
+    isInternalMCPServerOfName(arg.internalMCPServerId, "web_search_&_browse")
   );
 }
